+++ packages/demo-wallet-shared/methods/sep6/pollWithdrawUntilComplete.ts
@@ -1,16 +1,14 @@
 import {
-  Account,
   Asset,
   Keypair,
-  Operation,
   Horizon,
-  TransactionBuilder,
 } from "@stellar/stellar-sdk";
 import { log } from "../../helpers/log";
-import { createMemoFromType } from "../createMemoFromType";
 import { AnyObject, TransactionStatus } from "../../types/types";
-import { getNetworkConfig } from "../../helpers/getNetworkConfig";
-import { SmartWalletService } from "../../services/SmartWalletService";
+import {
+  sendFromClassicAccount,
+  sendFromContractAccount,
+} from "../sendWithrawPayment";
 
 export const pollWithdrawUntilComplete = async ({
   amount,
@@ -164,82 +162,3 @@ export const pollWithdrawUntilComplete = async ({
     requiredCustomerInfoUpdates,
   };
 };
-
-async function sendFromClassicAccount(
-  amount: string,
-  transactionJson: any,
-  keypair: Keypair,
-  server: Horizon.Server,
-  paymentAsset: Asset,
-  networkPassphrase: string,
-) {
-  const memo = createMemoFromType(
-    transactionJson.transaction.withdraw_memo,
-    transactionJson.transaction.withdraw_memo_type,
-  );
-
-  log.request({
-    title: "Fetching account sequence number",
-    body: keypair.publicKey(),
-  });
-
-  // eslint-disable-next-line no-await-in-loop
-  const { sequence } = await server
-    .accounts()
-    .accountId(keypair.publicKey())
-    .call();
-
-  log.response({
-    title: "Fetching account sequence number",
-    body: sequence,
-  });
-
-  const account = new Account(keypair.publicKey(), sequence);
-  const txn = new TransactionBuilder(account, {
-    fee: getNetworkConfig().baseFee,
-    networkPassphrase,
-  })
-    .addOperation(
-      Operation.payment({
-        destination:
-        transactionJson.transaction.withdraw_anchor_account,
-        asset: paymentAsset,
-        amount,
-      }),
-    )
-    .addMemo(memo)
-    .setTimeout(0)
-    .build();
-
-  txn.sign(keypair);
-
-  log.request({
-    title: "Submitting withdrawal transaction to Stellar",
-    body: txn,
-  });
-
-  // eslint-disable-next-line no-await-in-loop
-  return await server.submitTransaction(txn);
-}
-
-async function sendFromContractAccount (
-  asset: Asset,
-  fromAcc: string,
-  toAcc: string,
-  amount: string,
-  contractId: string,
-) {
-  const swService = SmartWalletService.getInstance();
-
-  log.request({
-    title: "Submitting withdrawal transaction to Stellar",
-  });
-
-  return await swService.transfer(
-    asset.contractId(getNetworkConfig().rpcNetwork),
-    fromAcc,
-    toAcc,
-    Number(amount),
-    contractId,
-  );
-}
